test(step3): path 패키지 분류
