+++ src/main/java/nextstep/member/ui/AuthenticationPrincipalArgumentResolver.java
@@ -3,6 +3,7 @@ package nextstep.member.ui;
 import nextstep.member.application.JwtTokenProvider;
 import nextstep.member.domain.AuthenticationPrincipal;
 import nextstep.member.domain.LoginMember;
+
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -14,28 +15,31 @@ import java.util.List;
 
 @Component
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
-    private JwtTokenProvider jwtTokenProvider;
+	private JwtTokenProvider jwtTokenProvider;
 
-    public AuthenticationPrincipalArgumentResolver(JwtTokenProvider jwtTokenProvider) {
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
+	public AuthenticationPrincipalArgumentResolver(JwtTokenProvider jwtTokenProvider) {
+		this.jwtTokenProvider = jwtTokenProvider;
+	}
 
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(AuthenticationPrincipal.class);
-    }
+	@Override
+	public boolean supportsParameter(MethodParameter parameter) {
+		return parameter.hasParameterAnnotation(AuthenticationPrincipal.class);
+	}
 
-    @Override
-    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
-        String authorization = webRequest.getHeader("Authorization");
-        if (!"bearer".equalsIgnoreCase(authorization.split(" ")[0])) {
-            throw new AuthenticationException();
-        }
-        String token = authorization.split(" ")[1];
+	@Override
+	public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
+		String authorization = webRequest.getHeader("Authorization");
+		if (authorization == null) {
+			return null;
+		}
+		if (!"bearer".equalsIgnoreCase(authorization.split(" ")[0])) {
+			throw new AuthenticationException();
+		}
+		String token = authorization.split(" ")[1];
 
-        Long id = Long.parseLong(jwtTokenProvider.getPrincipal(token));
-        List<String> roles = jwtTokenProvider.getRoles(token);
+		Long id = Long.parseLong(jwtTokenProvider.getPrincipal(token));
+		List<String> roles = jwtTokenProvider.getRoles(token);
 
-        return new LoginMember(id, roles);
-    }
+		return new LoginMember(id, roles);
+	}
 }
