+++ src/main/java/nextstep/subway/domain/SubwayMap.java
@@ -11,6 +11,7 @@ import java.util.stream.Collectors;
 
 import lombok.RequiredArgsConstructor;
 import nextstep.subway.applicaion.dto.SearchType;
+import nextstep.subway.path.Path;
 
 @RequiredArgsConstructor
 public class SubwayMap {
