+++ src/main/java/nextstep/subway/path/PathController.java
@@ -1,15 +1,12 @@
-package nextstep.subway.ui;
+package nextstep.subway.path;
 
 import lombok.RequiredArgsConstructor;
 import nextstep.member.domain.AuthenticationPrincipal;
 import nextstep.member.domain.LoginMember;
-import nextstep.subway.applicaion.PathService;
-import nextstep.subway.applicaion.dto.PathResponse;
-import nextstep.subway.applicaion.dto.SearchType;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -21,8 +18,10 @@ public class PathController {
 	@GetMapping("/paths")
 	public ResponseEntity<PathResponse> findPath(
 		@AuthenticationPrincipal LoginMember loginMember,
-		@RequestParam Long source, @RequestParam Long target,
-		@RequestParam(name = "type") SearchType type) {
-		return ResponseEntity.ok(pathService.findPath(loginMember, source, target, type));
+		@ModelAttribute PathRequest pathRequest) {
+		if (loginMember == null) {
+			return ResponseEntity.ok(pathService.findPathGuest(pathRequest));
+		}
+		return ResponseEntity.ok(pathService.findPathForMember(loginMember, pathRequest));
 	}
 }
