+++ src/main/java/nextstep/subway/path/PathRequest.java
@@ -1,0 +1,13 @@
+package nextstep.subway.path;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import nextstep.subway.applicaion.dto.SearchType;
+
+@Getter
+@AllArgsConstructor
+public class PathRequest {
+	private Long source;
+	private Long target;
+	private SearchType type;
+}
